pyproject.toml
[tool.poetry]
name = "noether-bot"
version = "1.0.0"
description = "Automated delta-neutral hedging bot for EulerSwap JIT liquidity pools"
authors = ["NoetherBot Team"]
readme = "README.md"
python = "^3.11"
packages = [
    { include = "swap_monitor" },
    { include = "strategy_engine" },
    { include = "risk_manager" },
    { include = "exchange_manager" },
    { include = "database_manager" },
    { include = "config_manager" },
    { include = "logger_manager" },
    { include = "models" },
    { include = "tui" }
]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.0"
ccxt = "^4.2.0"
web3 = "^6.15.0"
textual = "^0.47.0"
plotext = "^5.2.8"
aiohttp = "^3.9.1"
asyncio = "^3.4.3"
pydantic = "^2.5.3"
sqlalchemy = "^2.0.25"
alembic = "^1.13.1"
rich = "^13.7.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
black = "^23.12.1"
flake8 = "^7.0.0"
mypy = "^1.8.0"
isort = "^5.13.2"
pre-commit = "^3.6.0"
types-python-dateutil = "^2.8.19"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"